pyproject.toml
[tool.poetry]
name = "second-course-in-statistical-learning"
version = "0.1.0"
description = "Second Course in Statistical Learning"
authors = ["Pedro Albuquerque"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
numpy = "^2.2"
pandas = "^2.2"
matplotlib = "^3.10"
seaborn = "^0.13"
scikit-learn = "^1.6"
scipy = "^1.15"
statsmodels = "^0.14"
jupyterlab = "^4.3"
ipykernel = "^6.29"
playwright = "^1.50"
pyro-ppl = {version = ">=1.9.1", python = ">=3.13,<3.15"}
torch = {version = ">=2.10.0", python = ">=3.13,<3.15"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-cov = "^6.0"
ruff = "^0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
